Refactor Ridge route state and scene runtime
